Title: 두 수의 합, Time: 139ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 240119/두 수의 합/sum-of-two-num.py
@@ -14,4 +14,4 @@ for i, j in combinations(numbers, 2):
         dic[j] = i
         num += 1
 
-print(num/2)
+print(int(num/2))
